+++ QueryController.mm
@@ -3599,11 +3599,11 @@ extern "C"
                 if( days < 0) yearOld = @"";
                 else yearOld = [NSString stringWithFormat: NSLocalizedString( @"%d d", @"d = day"), days];
             }
-            else yearOld = [NSString stringWithFormat: @"%d%@", months, NSLocalizedString( @" m", @"m = month")];
+            else yearOld = [NSString stringWithFormat: @"%d%@", (int) months, NSLocalizedString( @" m", @"m = month")];
         }
-        else yearOld = [NSString stringWithFormat: @"%d%@ %d%@", years, NSLocalizedString( @" y", @"y = year"), months, NSLocalizedString( @" m", @"m = month")];
+        else yearOld = [NSString stringWithFormat: @"%d%@ %d%@", (int) years, NSLocalizedString( @" y", @"y = year"), months, NSLocalizedString( @" m", @"m = month")];
     }
-    else yearOld = [NSString stringWithFormat: @"%d%@", years, NSLocalizedString( @" y", @"y = year")];
+    else yearOld = [NSString stringWithFormat: @"%d%@", (int) years, NSLocalizedString( @" y", @"y = year")];
     
 	[yearOldBirth setStringValue: yearOld];
 }
